Rename: 마약과 마약 은어 패키지 분리
