+++ src/main/java/com/vigilante/retriever/v1/argot/domain/graphview/ArgotGraphView.java
@@ -1,6 +1,8 @@
-package com.vigilante.retriever.v1.drug.domain.graphview;
+package com.vigilante.retriever.v1.argot.domain.graphview;
 
 import java.util.Set;
+
+import com.vigilante.retriever.v1.drug.domain.graphview.DrugGraphView;
 
 import lombok.Builder;
