+++ src/main/java/com/vigilante/retriever/v1/channel/domain/graphview/ChannelGraphView.java
@@ -2,7 +2,7 @@ package com.vigilante.retriever.v1.channel.domain.graphview;
 
 import java.util.Set;
 
-import com.vigilante.retriever.v1.drug.domain.graphview.ArgotGraphView;
+import com.vigilante.retriever.v1.argot.domain.graphview.ArgotGraphView;
 
 import lombok.Builder;
